requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.9.3
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
asyncio==3.4.3
attrs==25.1.0
autobahn==19.11.2
Automat==24.8.1
aws-lambda-powertools==1.25.5
aws-xray-sdk==2.14.0
bcrypt==4.1.2
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
boto3==1.37.20
botocore==1.37.20
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dateutils==0.6.12
debugpy==1.8.12
dhanhq==2.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
fyers_apiv3==3.1.7
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
grpcio-tools==1.64.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==0.10.0
joblib==1.4.2
keras==3.9.0
kiteconnect==5.0.1
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
nsetools==1.0.11
numpy==2.1.3
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
passlib==1.7.4
pathspec==0.12.1
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-socketio==5.12.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.2
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
service-identity==24.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
starlette==0.45.3
sympy==1.13.1
ta==0.11.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
twilio==9.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
upstox==2.0.3
upstox-python-sdk==2.12.0
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.1
watchfiles==1.0.4
webencodings==0.5.1
websocket-client==1.6.1
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.52
zope.interface==7.2
